pyproject.toml
[project]
name = "langchainbasics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
    "wikipedia>=1.4.0",
]
